src/main/de/hsmainz/cs/semgis/arqextension/geometry/attribute/NDims.java
package de.hsmainz.cs.semgis.arqextension.geometry.attribute;

import org.apache.jena.datatypes.DatatypeFormatException;
import org.apache.jena.sparql.expr.ExprEvalException;
import org.apache.jena.sparql.expr.NodeValue;
import org.apache.jena.sparql.function.FunctionBase1;
import org.locationtech.jts.geom.Coordinate;
import org.locationtech.jts.geom.Geometry;

import io.github.galbiston.geosparql_jena.implementation.GeometryWrapper; 

public class NDims extends FunctionBase1 {

	@Override
	public NodeValue exec(NodeValue arg0) {
        try {
            GeometryWrapper geometry = GeometryWrapper.extract(arg0);
            Geometry geom = geometry.getXYGeometry();
            Coordinate coord=geom.getCoordinates()[0];
            if(!Double.isNaN(coord.getM()) || !Double.isNaN(coord.getZ())){
            	return NodeValue.makeInteger(3);
            }
            else{
            	return NodeValue.makeInteger(2);
            } 
        } catch (DatatypeFormatException ex) {
            throw new ExprEvalException(ex.getMessage(), ex);
        }
	}

}
